// src/types/form.ts
export interface AnswersState {
    question1: boolean | null;
    question2: boolean | null;
    question3: boolean | null;
    question4: boolean | null;
    question5: boolean | null;
    question6: boolean | null;
    question7: boolean | null;
    question8: boolean | null;
    question9: boolean | null;
    question10: boolean | null;
}